refactor(server): add server to separate backend and frontend logic
