+++ server/entities/animal.entity.ts
